Teste_de_Transformacao_de_Dados/requirements.txt
tabula-py>=2.3.0
pandas>=1.3.0
PyPDF2>=2.0.0
JPype1>=1.3.0